address editing deleting and updating policy

+++ app/Http/Controllers/UserAddressesController.php
@@ -37,6 +37,8 @@ class UserAddressesController extends Controller
 
     public function edit(UserAddress $user_address)
     {
+        $this->authorize('own', $user_address);
+
         return view('user_addresses.create_and_edit', ['address' => $user_address]);
     }
 
@@ -44,6 +46,8 @@ class UserAddressesController extends Controller
 
     public function update(UserAddress $user_address, UserAddressRequest $request)
     {
+        $this->authorize('own', $user_address);
+
         $user_address->update($request->only([
             'state',
             'city',
@@ -57,6 +61,9 @@ class UserAddressesController extends Controller
     }
 
     public function destroy(UserAddress $user_address){
+
+        $this->authorize('own', $user_address);
+
         $user_address->delete();
 //        return redirect()->route('user_addresses.index');
         return [];
